added set_debug if in exec script and secret_key generation code in install script

+++ exec.sh
@@ -2,6 +2,12 @@
 # install complete of management of dockers
 # containers and images
 
+export GEM_SET_DEBUG="True"
+if [ $GEM_SET_DEBUG ]; then
+    export PS4='+${BASH_SOURCE}:${LINENO}:${FUNCNAME[0]}: '
+    set -x
+fi
+
 export name_project="$1"
 export no_exec_test="$2"
 
@@ -9,7 +15,7 @@ if [ "$name_project" = "" ]; then
     echo "Need to add name project as argument"
     exit 1
 fi
-if [ -d "$name_project" ]; then
+if [ -d "$name_project" -a -f "${name_project}/.env" ]; then
     cd "$name_project"
     docker-compose down
     docker system prune
@@ -18,7 +24,6 @@ if [ -d "$name_project" ]; then
     
     cd -
 fi
-export GEM_SET_DEBUG="True"
 export GEM_GIT_PACKAGE="oq-platform3"
 export GEM_GIT_REPO="git@github.com:gem"
 export NO_EXEC_TEST=$no_exec_test
